# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API барои системаи тиҷоратии электронӣ",
        "description": "Санади API барои ҳама контроллерҳо",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/stats/orders-count": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Шумораи фармоишҳо",
                "description": "Ин endpoint шумораи фармоишҳои пӯшидашударо дар давраи интихобшуда бармегардонад",
                "operationId": "66eb974de5ed4d4dcb9e21ee1ae67c23",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Санаи оғоз (формат: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Санаи анҷом (формат: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Шумораи фармоишҳо",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "integer",
                                            "example": 120
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/stats/orders-sales-sum": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Ҷамъбасти фурӯш",
                "description": "Ин endpoint маблағи умумии фармоишҳои пӯшидашударо бармегардонад",
                "operationId": "1a6e4d3ebde369f92b620434222d5c4c",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Санаи оғоз",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Санаи анҷом",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Маблағи умумӣ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "number",
                                            "example": 5600.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/admin/stats/orders-count-by-days": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Шумораи фармоишҳо аз рӯи рӯз",
                "description": "Ин endpoint шумораи фармоишҳоро барои ҳар рӯз дар давраи интихобшуда бармегардонад",
                "operationId": "92436b954330458737a19906191bc5b5",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Санаи оғоз",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Санаи анҷом",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Рӯйхати рӯзҳо бо шумораи фармоишҳо",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2025-01-01"
                                                    },
                                                    "orders_count": {
                                                        "type": "integer",
                                                        "example": 15
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}